+++ frontend/src/features/missions/MissionForm/ActionForm/ControlForm/InfractionForm/InfractionForm.tsx
@@ -2,12 +2,12 @@ import { useField } from 'formik'
 import { Form, Button, ButtonToolbar } from 'rsuite'
 import styled from 'styled-components'
 
-import { COLORS } from '../../../constants/constants'
-import { infractionTypeEnum, formalNoticeEnum, actionTargetTypeEnum } from '../../../domain/entities/missions'
-import { FormikCheckbox } from '../../../uiMonitor/CustomFormikFields/FormikCheckbox'
-import { FormikInput } from '../../../uiMonitor/CustomFormikFields/FormikInput'
-import { FormikRadioGroup } from '../../../uiMonitor/CustomFormikFields/FormikRadioGroup'
-import { FormikTextarea } from '../../../uiMonitor/CustomFormikFields/FormikTextarea'
+import { COLORS } from '../../../../../../constants/constants'
+import { infractionTypeEnum, formalNoticeEnum, actionTargetTypeEnum } from '../../../../../../domain/entities/missions'
+import { FormikCheckbox } from '../../../../../../uiMonitor/CustomFormikFields/FormikCheckbox'
+import { FormikInput } from '../../../../../../uiMonitor/CustomFormikFields/FormikInput'
+import { FormikRadioGroup } from '../../../../../../uiMonitor/CustomFormikFields/FormikRadioGroup'
+import { FormikTextarea } from '../../../../../../uiMonitor/CustomFormikFields/FormikTextarea'
 import { InfractionFormHeaderCompany } from './InfractionFormHeaderCompany'
 import { InfractionFormHeaderVehicle } from './InfractionFormHeaderVehicle'
 import { NatinfSelector } from './NatinfSelector'
