+++ frontend/src/features/missions/MissionForm/ActionForm/NoteForm.tsx
@@ -2,11 +2,11 @@ import { useField } from 'formik'
 import { Form, IconButton } from 'rsuite'
 import styled from 'styled-components'
 
-import { COLORS } from '../../../constants/constants'
-import { actionTypeEnum } from '../../../domain/entities/missions'
-import { FormikTextarea } from '../../../uiMonitor/CustomFormikFields/FormikTextarea'
-import { ReactComponent as DeleteSVG } from '../../../uiMonitor/icons/Delete.svg'
-import { ReactComponent as NoteSVG } from '../../../uiMonitor/icons/Note_libre.svg'
+import { COLORS } from '../../../../constants/constants'
+import { actionTypeEnum } from '../../../../domain/entities/missions'
+import { FormikTextarea } from '../../../../uiMonitor/CustomFormikFields/FormikTextarea'
+import { ReactComponent as DeleteSVG } from '../../../../uiMonitor/icons/Delete.svg'
+import { ReactComponent as NoteSVG } from '../../../../uiMonitor/icons/Note_libre.svg'
 
 export function NoteForm({ currentActionIndex, remove, setCurrentActionIndex }) {
   const [actionTypeField] = useField(`envActions.${currentActionIndex}.actionType`)
